solid-ads: Align with react-ads structure
