+++ solid-ads/src/lib/googletag.ts
@@ -1,0 +1,10 @@
+const googleTagStub = {
+    display: () => {},
+    enableServices: () => {},
+    pubads: () => {},
+    cmd: [],
+  };
+  
+  const getGoogleTag = () => window.googletag || googleTagStub;
+
+  export { getGoogleTag };
